refactor(scripting-ui): addtype-erased property support to inspector
Adds the ability to display and edit properties of C# components within the inspector window.
This allows for dynamic inspection of components defined outside of the core engine.
It achieves this by creating a `TypeErasedProperty` class that handles the display of type-erased fields, using component descriptions from the coordinator.
Also fixes an issue with undo/redo actions.